+++ packages/core/src/interactivity/DataComponent.tsx
@@ -1,6 +1,7 @@
-import { createElement, MouseEvent, useCallback } from 'react'
+import { createElement, MouseEvent, useCallback, useState } from 'react'
 import { DataComponentProps, InteractivityProps } from './types'
-import { SvgElementVariantProps, WithId } from '../general'
+import { CssProps, SvgElementVariantProps, WithId } from '../general'
+import { clone, merge } from 'lodash'
 
 export const DataComponent = <
     DataSpec extends WithId,
@@ -9,27 +10,58 @@ export const DataComponent = <
     component,
     data,
     props,
-    onMouseEnter,
-    onMouseMove,
-    onMouseLeave,
-    onClick,
+    handlers,
+    modifiers,
 }: DataComponentProps<DataSpec, ComponentProps>) => {
+    const [componentStyle, setComponentStyle] = useState<CssProps | undefined>(props.style)
+    const [key, setKey] = useState(0)
+    const style = props.style
+
     const handleMouseEnter = useCallback(
-        (event: MouseEvent) => onMouseEnter?.(data, event),
-        [data, onMouseEnter]
+        (event: MouseEvent) => {
+            handlers?.onMouseEnter?.(data, event)
+            if (modifiers?.onMouseEnter) {
+                setComponentStyle(merge(clone(style), modifiers.onMouseEnter))
+                setKey(key => key + 1)
+            }
+        },
+        [data, handlers, modifiers]
     )
     const handleMouseMove = useCallback(
-        (event: MouseEvent) => onMouseMove?.(data, event),
-        [data, onMouseMove]
+        (event: MouseEvent) => {
+            handlers?.onMouseMove?.(data, event)
+            if (modifiers?.onMouseMove) {
+                setComponentStyle(merge(clone(style), modifiers.onMouseMove))
+                setKey(key => key + 1)
+            }
+        },
+        [data, handlers, modifiers]
     )
     const handleMouseLeave = useCallback(
-        (event: MouseEvent) => onMouseLeave?.(data, event),
-        [data, onMouseLeave]
+        (event: MouseEvent) => {
+            handlers?.onMouseLeave?.(data, event)
+            if (modifiers?.onMouseLeave) {
+                setComponentStyle(merge(clone(style), modifiers.onMouseLeave))
+                setKey(key => key + 1)
+            }
+        },
+        [data, handlers, modifiers]
     )
-    const handleClick = useCallback((event: MouseEvent) => onClick?.(data, event), [data, onClick])
+    const handleClick = useCallback(
+        (event: MouseEvent) => {
+            handlers?.onClick?.(data, event)
+            if (modifiers?.onClick) {
+                setComponentStyle(merge(clone(style), modifiers.onClick))
+                setKey(key => key + 1)
+            }
+        },
+        [data, handlers, modifiers]
+    )
 
     return createElement(component, {
         ...props,
+        key,
+        style: componentStyle,
         onMouseEnter: handleMouseEnter,
         onMouseMove: handleMouseMove,
         onMouseLeave: handleMouseLeave,